requirements.txt
Django
djangorestframework
markdown       
django-filter  
django-hashid-field
django-decorator-include